extract FundWalletTabsContent into separate component
